requirements.txt
asgiref==3.8.1 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.2.0 ; python_version >= "3.10" and python_version < "4.0"
django-filter==24.2 ; python_version >= "3.10" and python_version < "4.0"
django==4.2.11 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.10" and python_version < "4.0"
drf-spectacular==0.27.2 ; python_version >= "3.10" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema-specifications==2023.12.1 ; python_version >= "3.10" and python_version < "4.0"
jsonschema==4.21.1 ; python_version >= "3.10" and python_version < "4.0"
psycopg2==2.9.9 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "4.0"
referencing==0.35.0 ; python_version >= "3.10" and python_version < "4.0"
rpds-py==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.11.0 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2024.1 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
uwsgi==2.0.23 ; python_version >= "3.10" and python_version < "4.0"
freezegun==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
